plan增加是否完成字段

+++ src/main/java/com/lixxy/depositplan/model/SaveRecordBean.java
@@ -9,7 +9,7 @@ public class SaveRecordBean {
 
     private Integer userId;
 
-    private Long money;
+    private Double money;
 
     private Integer isSaved;
 
@@ -41,11 +41,11 @@ public class SaveRecordBean {
         this.userId = userId;
     }
 
-    public Long getMoney() {
+    public Double getMoney() {
         return money;
     }
 
-    public void setMoney(Long money) {
+    public void setMoney(Double money) {
         this.money = money;
     }
